+++ src/Components/Parsers/IndexHints.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhpMyAdmin\SqlParser\Components\Lists;
+namespace PhpMyAdmin\SqlParser\Components\Parsers;
 
 use PhpMyAdmin\SqlParser\Components\IndexHint;
 use PhpMyAdmin\SqlParser\Parseable;
