+++ src/Statements/DeleteStatement.php
@@ -9,12 +9,12 @@ use PhpMyAdmin\SqlParser\Components\Condition;
 use PhpMyAdmin\SqlParser\Components\Expression;
 use PhpMyAdmin\SqlParser\Components\JoinKeyword;
 use PhpMyAdmin\SqlParser\Components\Limit;
-use PhpMyAdmin\SqlParser\Components\Lists\Conditions;
-use PhpMyAdmin\SqlParser\Components\Lists\ExpressionArray;
-use PhpMyAdmin\SqlParser\Components\Lists\JoinKeywords;
-use PhpMyAdmin\SqlParser\Components\Lists\OrderKeywords;
 use PhpMyAdmin\SqlParser\Components\OptionsArray;
 use PhpMyAdmin\SqlParser\Components\OrderKeyword;
+use PhpMyAdmin\SqlParser\Components\Parsers\Conditions;
+use PhpMyAdmin\SqlParser\Components\Parsers\ExpressionArray;
+use PhpMyAdmin\SqlParser\Components\Parsers\JoinKeywords;
+use PhpMyAdmin\SqlParser\Components\Parsers\OrderKeywords;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Statement;
 use PhpMyAdmin\SqlParser\TokensList;
