+++ tests/Components/ExpressionArrayTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\Lists\ExpressionArray;
+use PhpMyAdmin\SqlParser\Components\Parsers\ExpressionArray;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
