+++ tests/Components/JoinKeywordTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\Lists\JoinKeywords;
+use PhpMyAdmin\SqlParser\Components\Parsers\JoinKeywords;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
